A lot of Changes:
First of all I changed the sign of the force summation in the Evolution
part of the MD to a positive (as it should have been all along). For the
physics to stay right I had to change the sign of all the forces in the
springs and the bending forces. I forgot to change the signe of the
damping coefficient and that resulted in a lot of error. It is fixed
now.

I completed the membrane config file reader.

When a mesh file is imported the check() function will check all the
node distances. If the the max and min distances are such that they
might end up in a line and break down as a triangle, it will add a MD
step in between to fix the max bond distance cut off. The trajectory of
the pre MD calculations will be saved in a xyz file.

After the pre check is complete the MD will carry out normaly and with a
thermostat.