I have modefied the membrane bond length check process defined in the 'check' function:
1-The bond length adjustment will now begin with the bending potential
included + a very weak damper.

2-I wrote another potential called the relaxation potential which is a
FENE potential that includes a hookian cutt off potential as well. The
max and Min bond length are also set for a 9 percent larger values to
garantee that we will not run into trouble when using a regular
potential.

3- After a couple of thousand MD steps the strong damper kicks in and
smoothens out the membrane. Also this process drastically decreases the
energy potential and kinetic energy of the system. Hence, almost all the
nodes are in the 'free zone' region of the potential.

I have also modefied the FENE potential so that the programme terminates
if the bond distances exceed the MAX and MIN specified distances.