Hoda and I talked some changes she made to make the relaxation of membranes a part of the main.cpp.
My config files do not work right now. I am just going to commit this
version and pull the solution that Hoda has represented.

In the current version:

The Membrane is imported successfully
... Actin    .. ........ ............
The shared nodes are identified.

But as soon as the simulation begins, I gat 'nan' in hte trajectory.